If page is not found 404 is thrown correctly

+++ database/migrations/2020_05_07_152643_create_item_table.php
@@ -29,7 +29,7 @@ class CreateItemTable extends Migration
             $table->string('subCategorie')->nullable($value=true);
             $table->enum('colors', ['Red', 'Blue', 'Purple']);
             $table->string('dimensions');
-            $table->decimal('OverAllrating')->default(0);
+            $table->integer('OverAllrating')->default(0);
         });
     }
